Refactor ISSN-registry publisher list and item views
